Create Function to sort array of 0s,1s,2s
The program performs in place swapping of 0s 1s and 2s to give sorted output. If one is found the program increments mid and does not perform swap. If 0 is found it is swapped to the start of the array and 2 is swapped to the end and accordingly mid is incremented or high is decremented.